refactor: Program.cs dependency refactored to be more understandable
