+++ src/Modmail.NET.Web.Blazor/Dependency/ValidatorDependency.cs
@@ -1,0 +1,10 @@
+using FluentValidation;
+
+namespace Modmail.NET.Web.Blazor.Dependency;
+
+public static class ValidatorDependency
+{
+  public static void Configure(WebApplicationBuilder builder) {
+    builder.Services.AddValidatorsFromAssemblyContaining(typeof(ModmailBotProjectMarker));
+  }
+}
