+++ src/Modmail.NET.Web.Blazor/Modmail.NET.Web.Blazor.csproj
@@ -38,6 +38,7 @@
     <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="9.0.3" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="9.0.3" />
     <PackageReference Include="Radzen.Blazor" Version="6.2.9" />
+    <PackageReference Include="Scrutor" Version="6.0.1" />
     <PackageReference Include="Serilog" Version="4.2.0" />
     <PackageReference Include="Serilog.AspNetCore" Version="9.0.0" />
     <PackageReference Include="Serilog.Enrichers.Environment" Version="3.0.1" />
